Moved remote connection specifc calls to a separate module
