+++ src/remote_storage/mod.rs
@@ -1,2 +1,3 @@
 pub mod callback_service;
+pub mod connection_entry;
 pub mod storage_service;
